Move provider to storage package
